Fix failing fuzz test
It turns out that testing::StrCaseEq() is a recursive function and
this causes the stack to grow significantly during fuzz testing. This
test was not entirely correct in that the ASCII part of the string
should be checked with full case sensitivity.

In the CI action, space the -Ds consistently. Explicitly set URI_CXX17
for each job.

Add ranges support to punycode::encode()